requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
pathspec==0.10.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg==3.1
psycopg-binary==3.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.2
